requirements.txt
# Azure Functions Core
azure-functions

# HTTP requests
requests

# Database connectivity  
pyodbc